refactor(Home): refactor pages/Home folder
