+++ src/pages/Home/constants.js
@@ -1,0 +1,4 @@
+export const searchMode = {
+  NAME: 'search_by_name',
+  TYPE: 'search_by_type'
+}
